requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Data validation
pydantic>=2.7.2

# Google APIs
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0

# MCP
mcp>=1.1.0

# LangChain (for Phase 2)
langchain>=0.1.0
langchain-openai>=0.0.5

# Utilities
requests>=2.31.0